requirements.txt
fastapi
uvicorn
sqlmodel
firebase-admin
python-dotenv
jinja2
pandas
pdfplumber
backend



